Fix methods NewsServiceImpl

+++ src/main/java/megalab/service/impl/UserServiceImpl.java
@@ -29,11 +29,11 @@ import java.time.ZonedDateTime;
 @Transactional
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
-    private  final UserRepository userRepository;
+    private final UserRepository userRepository;
     private final UserInfoRepository userInfoRepository;
     private final PasswordEncoder passwordEncoder;
 
-    private User getAuthentication(){
+    private User getAuthentication() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String name = authentication.getName();
         return userRepository.getUserByUserInfoNickName(name).orElseThrow(() -> new NotFoundException("User with nickName: " + name + " us bit found!"));
@@ -45,7 +45,7 @@ public class UserServiceImpl implements UserService {
         Page<UserResponse> allUsers = userRepository.getAllUsers(pageable);
         return UserPagination.builder()
                 .userResponses(allUsers.getContent())
-                .currentPage(allUsers.getNumber()+1)
+                .currentPage(allUsers.getNumber() + 1)
                 .pageSize(allUsers.getTotalPages())
                 .build();
     }
@@ -99,7 +99,7 @@ public class UserServiceImpl implements UserService {
         userInfo.setModifiedAt(ZonedDateTime.now());
 
 
-        if (authentication.getUserInfo().getRole().equals(Role.ADMIN)){
+        if (authentication.getUserInfo().getRole().equals(Role.ADMIN)) {
             userRepository.save(user);
             userInfoRepository.save(userInfo);
             return SimpleResponse.builder()
@@ -107,14 +107,14 @@ public class UserServiceImpl implements UserService {
                     .message(String.format("User with id:%s successfully updated", userId))
                     .build();
         }
-        if (user.equals(authentication)){
+        if (user.equals(authentication)) {
             userRepository.save(user);
             userInfoRepository.save(userInfo);
             return SimpleResponse.builder()
                     .status(HttpStatus.OK)
                     .message(String.format("User with id:%s successfully updated", userId))
                     .build();
-        }else throw new BadRequestException("You can not update this user!");
+        } else throw new BadRequestException("You can not update this user!");
 
     }
 
@@ -123,7 +123,7 @@ public class UserServiceImpl implements UserService {
         User user = userRepository.findById(id).orElseThrow(() -> new NotFoundException(String.format("User with id:%s is not present", id)));
         UserInfo userInfo = userInfoRepository.findById(user.getUserInfo().getId()).orElseThrow(() -> new NotFoundException(String.format("UserInfo with id:%s is not present", user.getUserInfo().getUser())));
         User user1 = getAuthentication();
-        if (user1.getUserInfo().getRole().equals(Role.ADMIN)){
+        if (user1.getUserInfo().getRole().equals(Role.ADMIN)) {
             return UserResponse.builder()
                     .id(user.getId())
                     .firstName(user.getFirstName())
@@ -133,7 +133,7 @@ public class UserServiceImpl implements UserService {
                     .nickName(userInfo.getNickName())
                     .build();
         }
-        if (user.equals(user1)){
+        if (user.equals(user1)) {
             return UserResponse.builder()
                     .id(user.getId())
                     .firstName(user.getFirstName())
@@ -142,7 +142,7 @@ public class UserServiceImpl implements UserService {
                     .image(user.getImage())
                     .nickName(userInfo.getNickName())
                     .build();
-        }else throw new BadRequestException("You can not get this user!");
+        } else throw new BadRequestException("You can not get this user!");
 
     }
 
@@ -153,6 +153,12 @@ public class UserServiceImpl implements UserService {
 
         UserInfo userInfo = userInfoRepository.findById(user.getUserInfo().getId()).orElseThrow(() -> new NotFoundException(String.format("UserInfo with id:%s is not present", user.getUserInfo().getUser())));
         if (authentication.getUserInfo().getRole().equals(Role.ADMIN)) {
+
+
+            if (user.getUserInfo().getRole().equals(Role.ADMIN)) {
+
+
+
                 userRepository.deleteById(id);
                 return SimpleResponse.builder()
                         .status(HttpStatus.OK)
@@ -168,11 +174,19 @@ public class UserServiceImpl implements UserService {
             } else throw new BadRequestException("You can not delete this account");
 
 
+
         }
 
-    @Override
-    public UserPagination searchUser(String word, int currentPage, int pageSize) {
+
+        }
         return null;
     }
 }
+
+
+//    @Override
+//    public UserPagination searchUser(String word, int currentPage, int pageSize) {
+//        return null;
+//    }
+
